+++ Source/LogBridge.EnterpriseLibrary/EnterpriseLibraryWrapper.cs
@@ -69,7 +69,6 @@ namespace SoftwarePassion.LogBridge.EnterpriseLibrary
                 etlProperties[LogConstants.CorrelationIdKey] = null;
             etlProperties[LogConstants.EventIdKey] = logData.EventId;
             etlProperties[LogConstants.MachineNameKey] = logData.MachineName;
-            etlProperties[LogConstants.ApplicationNameKey] = logData.ApplicationName;
             etlProperties[LogConstants.ProcessNameKey] = logData.ProcessName;
             etlProperties[LogConstants.FilenameKey] = logData.LogLocation.FileName;
             etlProperties[LogConstants.LineNumberKey] = logData.LogLocation.LineNumber;