+++ Source/LogBridge.Log4Net/Log4NetWrapper.cs
@@ -38,7 +38,8 @@ namespace SoftwarePassion.LogBridge.Log4Net
                 LocationInfo = ToLog4NetLocationInfo(logData.LogLocation),
                 UserName = logData.Username,
                 Domain = logData.AppDomainName,                
-                Properties = ToLog4NetProperties(logData),
+                Properties = 
+                ToLog4NetProperties(logData),
                 LoggerName = activeLogger.Name,
                 ExceptionString = logData.Exception == null 
                     ? string.Empty
@@ -124,7 +125,6 @@ namespace SoftwarePassion.LogBridge.Log4Net
 
             log4NetProperties[LogConstants.EventIdKey] = logData.EventId;
             log4NetProperties[LogConstants.MachineNameKey] = logData.MachineName;
-            log4NetProperties[LogConstants.ApplicationNameKey] = logData.ApplicationName;
             log4NetProperties[LogConstants.ProcessNameKey] = logData.ProcessName;
             log4NetProperties[LogConstants.ExceptionKey] = logData.Exception;
             return log4NetProperties;